product variant item

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -92,10 +93,10 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $varinat = ProductVariant::findOrFail($id);
-        // $variantItemCheck = ProductVariantItem::where('product_variant_id', $varinat->id)->count();
-        // if($variantItemCheck > 0){
-        //     return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
-        // }
+        $variantItemCheck = ProductVariantItem::where('product_variant_id', $varinat->id)->count();
+        if($variantItemCheck > 0){
+            return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
+        }
         $varinat->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
